fix #111 and log errors during watch
